split router and controller
